Daily build for Wed Feb  1 23:40:48 EET 2012
Added framework for handling bond order descriptors in the core.

Instead of defining the bond order factors just in Python, I decided to
copy the system used for potentials and define the characteristics
of the bond order factors in the core. This update includes the inquiry
subroutines for accessing the information contained in the core.
PyInterface.f90 has not been changed so there is no access to these functions
yet.

This will also mean that the previous update on Python side will have to
be altered quite a lot to fit this scheme. The great benefit is that the
new method will allow for easy inclusion of any new bond-order factors.