Daily build for Thu Feb  9 22:59:30 EET 2012
Continued building framework for handling bond order descriptors in the core.

The routines for inquiring the core and calculating individual terms of the
bond order factors are in place, but untested. PyInterface.f90 has the method
for calling the core, but the subroutine in the core where one loops over particles
(core_calculate_bond_orders) is still unfinished. The inquiry methods are missing and
the Python side does not connect to Fortran yet.