Daily build for Mon Feb 13 23:16:50 EET 2012
Finished building framework for handling bond order descriptors in the core.

The routines for inquiring the core and calculating individual terms of the
bond order factors are in place, but untested. This includes PyInterface.f90.
Importing pysic in python now automatically initializes the bond order factor
descriptors in the core so that they can be inquired by pysic_fortran.

The general methods in pysic for accessing the interface are being implemented.
The aim is to have most of the potential inquiry functions to recognize if its
argument is a bond order factor and then target them instead for similar
functionality.