Daily build for Tue Feb 14 22:42:00 EET 2012
Wrote the framework for handling BondOrderParameters in Python.

Now the BondOrderParameters objects can be manipulated in a
framework very similar to that of Potential objects. Each parameter
object has cutoffs, a set of parameters for elements/pairs/etc.
and a list of element symbols to target.

Next, the Coordinator needs to be rewritten so that it can handle these
parameters objects. Then, the Coordinator has to be passed to Potentials.
Finally, the translation of Python parameters to the core needs to be
implemented. The Fortran routines do already exist so it is fairly straightforward.