Daily build for Wed Feb 15 23:04:05 EET 2012
Added framework for creating the bond order factors in the core.

In addition, the way the bond factor evaluation loops was changed to be similar to how potential evaluation works. This means that the evaluation of the actual bond factor terms must be modified.

At this point nothing triggers the calculation of bond orders. It needs to happen when calculating energies or forces.