Daily build for Fri Feb 17 17:08:40 EET 2012
Bond order factors are now passed to the core with potentials that contain them. The potentials and bond factors are assigned indices to ensure the correct factors are joint with the correct potentials. In addition, the coordinator objects in pysic are given the same factors to allow the calculation of bond order factors separately from potential evaluation. There is no checkup for the validity of the core when doing this.

The bond order factors are not yet applied to the potentials during force evaluation even if the potential has bond order factors associated with it. However, it is possible to trigger the bond order calculation manually by using the Coordinator object. This is convenient for debugging and testing. So far at least the neighbor counting factor works as intended. More rigorous testing needs to be carried out still.

The evaluation of cutoffs have been updated to take into account the bond order factors present. So, even if a particular atom is not experiencing a potential, it might be contributing to bond orders and must appear in the correct neighbor lists.