Daily build for Sat Feb 25 01:19:44 EET 2012
Multiplicative bond order factors are now fully working.

Here, multiplicative refers to factors that one can scale the potential energy with. The current scheme allows one to define a bond order factor and attach it to any potential. Comb will also require a potential which takes coordination as an argument, and this is not yet implemented. However, that should be relatively painless to implement within the current framework and will be addressed later.

The calculation of forces and energies has been tested for 1-3 body interactions and bond factors in a serial environment. More elaborate testing is still needed, but so far all the functionality seems to work. Some things to test further include having multiple potentials and bond factors present. The MPI framework needs to be rechecked since some of the bond order functions are called from within an MPI parallelized loop while some are not. Therefore one needs to be able to choose whether to run the routines in parallel or not, on the fly.

All around tidying and commenting of the code is also needed. Especially the Sphinx documentation is downrigth wrong, since it now explains things that were scrapped long ago (and the current implementation is more elegant anyway). After this, the branch will be merged with the master version.