Daily build for Tue Mar 13 23:01:35 EET 2012
Added plenty of comments in the code, especially in the Fortran side which, although written cleanly and using relatively good programming practices, clearly did not have adequate explanations on the logic behind some of the routines. In addition, I wrote a script for automatically generating Sphinx input files from the Fortran source code and comments. Due to this, the documentation now includes descriptions of all the Fortran modules as well as the Python ones. Although the Fortran routines are meant to be hidden from the user, this additional documentation is important for development purposes.

There are still Fortran modules without the formatted comments the documentation script parses, so some of the files in the html documentation appear rather empty. The names of subroutines and global variables are parsed even without the comments, though.