Implemented charge relaxation.
Basic routines have been added to communicate the atomic charges between the Fortran and Python sides.

In Python, a new class called ChargeRelaxation has been implemented to handle the relaxation of charges. The class is similar in interface to Potential and BondOrderParameters - it can be given a type and a set of parameters - but its data is not stored in the core. Therefore, no elaborate routines are required to communicate the descriptions of the types from Fortran to Python.
Currently only damped dynamic relaxation of charges is available, but the framework to easily add more types is available. The damped charge dynamics are currently implemented in Python. Likely not much efficiency can be gained by implementing the routines in Fortran, but it should be explored nonetheless.

In Fortran, a charge dependent potential 'exponential' has been added. Also a routine for calculating electronegativities has been added, similar to that for calculating forces. These routines are accessed from the Pysic calculator. The electronegativities are, in effect, the forces guiding the charge relaxation and their
calculation is the computationally intensive task of the procedure. Hence the implementation in Fortran (with MPI).