First version of Ewald summation implemented.
To enable Ewald, the CoulombSummation class has been introduced in the Python interface. Passing an instance to the calculator switches on the Coulomb interaction.

The documentation has some brief description of this but it's still very crude lacking real instructions on how to use it.

There is Ewald energy and force calculation but no electronegativity evaluation yet!

Also fixed a bug in par finding when periodic images of the same atom should interact due to very long cutoffs (see function 'pick' in Geometry.f90).