Combined energy, force, and electronegativity calculation in a single routine in Core.f90.
There are still the same headers for calling the evaluations ('core_evaluate_energy' etc.) but these all now call the routine 'core_loop_over_local_interactions'. The reason for this rearrangement is that all the routines used to do the same loop over atom singles, doubles, and triples. This was a long piece of copy pasted code that is now only included only once. (It is still repeated for bond order factors, in a slightly different form. Those loops remain to be integrated as one.) Now, to add quadruplets (needed for dihedral angles) only this one loop has to be appended.