Performance boost in Ewald summation.
In force evaluation using Ewald summation a huge array was created for storing values that were not used frequently nor calculated in parallel. This resulted in lots of wasted memory and a massive performance bottleneck when the array was initialized. The issue was fixed simply by removing the array and evaluating the needed values on the fly.