Implemented stress calculation.
The stress tensor is now automatically evaluated when the forces are. This adds small additional cost to evaluating the forces, but since the full force evaluation is required to obtain the stress consistently in a system with periodic boundaries, this saves a lot of effort when the stress is needed. Especially, if one wishes to run NPT simulations, the stress tensor will be needed frequently.

The obtained stress values seem reasonable and agree with the calculated energies, but the routine has not been yet properly tested.