Major code restructuring as a Python package
The amount of code on the Python side was getting too large to keep
everything in a single file, so I restructured the code in smaller
pieces divided in a folder structure. Mostly the syntax seen by the
user is the same: import pysic gives access to pysic.Pysic,
pysic.Potential etc. but behind the scenes, for instance the Potential
class is now pysic.interactions.local.Potential.