Added a readme and a makefile (closes #6)
In order to make compiling easy, a Makefile is provided. This is
especially nice now with the new tidy package structure since compiling
with the make script automatically takes care of putting the compiled
core binary (pysic_fortran.so) in the right place.

Note that some things here and there may be slightly out of place in
the documentation now that the package structure has changed. The doc
evolves constantly though, so this is not a big deal and will be fixed
in the future.