Added tabulated pair potentials #3
A tabulated potential has been implemented. The potential reads from a
file a table of values for the potential and its derivative (negative
force) on a uniform grid of arbitrary range and density. The values are
then interpolated using cubic splines. The potential is still
undocumented and similar tabulated functions should be implemented also
for bond order summation and scaling.