Added the local charge self energy potential (closes #7)
Added a power law potential for the self energy of atomic charges. This
potential is quite important since also constrains the local charges.
With no constraints, you could end up in a situation where the charge
on one atom goes to infinity and the charge on another to negative
infinity (which is of course nonsense).