Created an interface for multiplied potentials #11
I'm testing out a framework where one can define local potentials as
products of elemental potentials. That is,

    V = sum v1*v2*...

A class MultipliedPotential has been added to allow for the definition
of such potentials and the core has been adjusted to allow for these
types of potential groups. The actual calculation of energies and
forces does not yet evaluate the products, though.

The benefit of being able to multiply potentials is the ability to take
a group of fairly general potentials and combine them to a broad
variety of more specialized interactions without having to hardcode all
the possibilities. For instance, spatial and charge dependencies can be
separated in their own 'elemental' potentials.