Implemented energy, force and e-negativity calculation of product potentials #11
Implemented the actual evaluation of physical quantities for multiplied
potentials. In essence, a multiplied potential is of the form

    V = sum v1*v2*...vn

and the corresponding force is

    F = - nabla V = sum (-nabla v1)*v2*...*vn +  ... + v1*...*(-nabla
vn).

The energy and force components v and -nabla v are calculated by the
already implemented 'regular' potentials.
The electronegativity is obtained similarly to the force.

Renamed the class handling the multiplication in Python to
ProductPotential.

Documentation is not yet up for this feature.