Added separate spatial and charge dependent exponential potentials (closes #8)
Separated the old exponential potential to spatial and charged
contributions. The old potential can be restored by using the
ProductPotential, but the new spatial part can be combined with other
potentials. Especially, it can be connected to the power decay
potential to construct the electrostatic potentials between Slater
orbitals (together with CoulombSummation).