Made the calculator return copies of the calculated arrays
Pysic used to return the calculated forces etc. as they were. However,
if ASE is run with constraints, the constraints may modify the forces
in place thus changing also the arrays stored in Pysic. So, if you did
a constrained calculation and then wanted to access the raw
unconstrained forces, that was not possible without a recalculation.
Now, since copies are passed on to ASE, the real forces are available
as they should.