Stability fixes (closes #21)
Fixed some issues where two atoms on top of each other would lead to
NaN energies. It is still possible to make the system explode by
putting atoms with a singular interaction on top of each other, but
there are cases (e.g., springs) that should handle being superposed.

Also fixed a bug in neighbor list generation. If cutoffs are longer
than cell size, the FastNeighborList will fail and thus Pysic falls
back on the ASE neighbor list. The switch was determined based on
maximum cutoff, but the list generating is done by cutoff + neighbor
list skin. So, itf your minimum cell size happened to be between max
cutoff and max cutoff + skin, Pysic tried to use FastNeighborList,
which got stuck because the cutoff was too long for making the space
partitioning.

Finally, there was a bug in monitoring the core when constraints were
applied, which would lead to Pysic thinking the core was not up to date
since the constrained and unconstrained data did not match.