Added absolute valued charge potential (closes #9)
Added a pair potential for absolute charges

    V(q1,q2) = sqrt(B1*B2)
    Bi = ai + bi * abs(qi - Qi)**ni

combined with the exponential potentials, this should allow one to
create the charge depenedent attractive and repulsive potentials of
COMB formalism.

This also marks the implementation of the base COMB potential.
Unfortunately, there are no wrappers for complicated potentials in
place yet, so one must build the potential piece by piece. Also the
polarization including components of later revisions of the potential
are not available (as there is no polarization implemented).