Implemented a framework for compound potentials
Often one wants to use a complicated potentials that are not directly
available in the core, but can be built by combining the various terms
that are. A framework has been implemented for wrapping such compound
potentials in a single class in the Python interface.

The wrapper class is called CompoundPotential, and it inherits most of
the interface from Potential. This class defines an "empty" potential,
and to define true potentials, one should write a new subclass of
CompoundPotential. One real potential, the Sutton-Chen potential, is
implemented here as the class SuttonChenPotential. Later, at least COMB
and ReaxFF should be included as well.

The framework is not polished yet. There is no documentation, many
convenience functions remain to be added, and the interface for adding
the potential to a calculator is different to normal Potentials (not
calc.add_potential(pot) but pot.build(calc)). These will hopefully be
fixed soon.