Bond order factor duplicate check
Added a duplicate check for bond order factors. When the bond order
factors are created in the Fortran core, every potential used to have
its own factor even if they actually had the same ones attached to them
in the Python side.

Now, the code checks for reuse of the same Coordinator when the core is
created, and if the same one appears with several potentials, it is
attached to them all without creating several copies. The immediate
benefit is that the code does not need to evaluate the same factors
several times during energy and force evaluation (as the factors are
precalculated at the start of evaluation).