Makefile corrections
Some fixes to makefile:

- compiling in modes other than 'parallel' created pysic with a suffix,
e.g., pysic_serial. This was incompatible with the new submodule
structure. Now all versions created with make are named pysic.
- Serial compilation had the noopt flag on, which led to the
compilation of a nonoptimized version. This is now fixed.

This commit also includes a bug fix in ChargeRelaxation where a routine
that was copied from CoulombSummation actually still referred to
CoulombSummation, not ChargeRelaxation.