Included 2pi factor in Ewald parameter determination
When converting from k_cutoff to k-space summation ranges in Ewald
setup, the ase.Atoms.get_reciprocal_cell() function is used which does
not include the 2pi factor usually used for the reciprocal cell. This
makes the k-cutoff essentially 2pi times longer than expected if the
standard definition of a reciprocal cell is assumed. This update adds
the factor 2pi in explicitly.