Apply k-space cutoff directly in Ewald
Previously the given k-cutoff radius was only used for determining the
k-summation cutoffs for Ewald, which decide how many times the sum goes
over shifted periodic images of the reciprocal cell. However, there is
really no need to calculate the structure factors for all k in that
region - which has a rectangular shape (or in general triclinic) - in
fact it makes more sense to use a spherical cutoff. Now the summation
over k is also truncated according to the absolute length of k to save
calculating all the structure factors.

Also a function for guessing Ewald parameters was included and a bug in
initializing the parameters upon creating a CoulombSummation was fixed.