Ewald combined with real space sum of potentials
The ewald summation is now done along with the rest of the real space
summation of potentials to avoid looping over neighbors twice. It is
possible to do the real space sum also along with the reciprocal sum in
the ewald subroutines, but when called from the force and energy
calculation routines, this is avoided.

Also fixed bugs that would lead to segmentation faults when a
calculation was run without any potentials. This was due to some arrays
being initialized only when potentials were being created.